chore: move generated types to other directory
