+++ packages/front/codegen.ts
@@ -5,14 +5,14 @@ const config: CodegenConfig = {
   schema: ['../backend/src/server/graphql/schema.graphql'],
   documents: 'src/**/*.graphql',
   generates: {
-    'src/types.ts': {
+    'src/graphql/types.ts': {
       plugins: ['typescript']
     },
     'src/': {
       preset: 'near-operation-file',
       presetConfig: {
         extension: '.generated.tsx',
-        baseTypesPath: 'types.ts'
+        baseTypesPath: 'graphql/types.ts'
       },
       plugins: ['typescript-operations', 'typescript-react-apollo']
     }
