+++ packages/front/src/graphql/types.ts
@@ -1,4 +1,3 @@
-/** Auto generated by graphql codegen */
 export type Maybe<T> = T | null;
 export type InputMaybe<T> = Maybe<T>;
 export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
